Revert "Fixed comparison on 64 bit system"
This reverts commit 65890a73d53a25b0809b22e41cdee19e3f4aa68d.

+++ media/libstagefright/SampleTable.cpp
@@ -246,7 +246,7 @@ status_t SampleTable::setSampleToChunkParams(
     for (uint32_t i = 0; i < mNumSampleToChunkOffsets; ++i) {
         uint8_t buffer[12];
 
-        if ((off64_t)(INT64_MAX - 8 - (i * 12)) < mSampleToChunkOffset) {
+        if ((off64_t)(SIZE_MAX - 8 - (i * 12)) < mSampleToChunkOffset) {
             return ERROR_MALFORMED;
         }
